requirements.txt
dj-database-url==0.5.0
Django==3.2.10
django-bootstrap4==3.0.1
django-crispy-forms==1.13.0
django-filter==2.4.0
django-google-maps==0.12.4
django-rest-framework==0.1.0
django-tables2==2.4.1
djangorestframework==3.12.4
folium==0.12.1.post1
geocoder==1.38.1
gunicorn==20.1.0
MarkupSafe==1.1.1
Markdown==3.3.4
nose==1.3.7
oauthlib==3.1.1
paynow==1.0.4
PyMySQL==1.0.2
sendgrid==3.6.5
sendgrid-django==4.2.0
